refactor: prepared branch for code review
