+++ package.json
@@ -21,6 +21,7 @@
   },
   "devDependencies": {
     "@evilmartians/lefthook": "^1.1.1",
-    "lerna": "^5.4.3"
+    "lerna": "^5.4.3",
+    "sass": "^1.62.1"
   }
 }
