+++ packages/client/package.json
@@ -16,7 +16,8 @@
     "eslint-config-prettier": "^8.5.0",
     "prettier": "^2.7.1",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "node-sass": "8.0.0"
   },
   "devDependencies": {
     "@testing-library/react": "^13.3.0",
