--- packages/client/src/App.css
@@ -1,26 +1,0 @@
-@font-face {
-    font-family: FutureMillennium;
-    src: url("./styles/fonts/FutureMillennium Black.ttf") format("truetype");
-    font-weight: 800;
-    font-style: normal;
-}
-
-@font-face {
-    font-family: FutureMillennium;
-    src: url("./styles/fonts/FutureMillennium.ttf") format("truetype");
-    font-weight: 400;
-    font-style: normal;
-}
-
-body {
-    width: 100vw;
-    height: 100vh;
-    overflow: hidden;
-    padding: 0;
-    margin: 0;
-}
-
-#root {
-    width: 100%;
-    height: 100%;
-}
