+++ packages/client/src/App.scss
@@ -1,0 +1,32 @@
+@import './src/styles/vars';
+
+html {
+  font-family: $font-family;
+  font-weight: 400;
+  font-size: 20px;
+}
+
+body {
+  width: 100vw;
+  height: 100vh;
+  overflow: hidden;
+  padding: 0;
+  margin: 0;
+}
+
+input {
+  outline: none;
+}
+
+a {
+  text-decoration: none;
+  color: $font-color;
+
+}
+
+#root {
+  display: flex;
+  width: 100%;
+  height: 100%;
+
+}
