+++ packages/client/src/assets/bg.jpg
