+++ packages/client/src/components/HOC/Layout/style.module.scss
@@ -1,0 +1,4 @@
+.Layout {
+    display: flex;
+    flex-grow: 1;
+}
