+++ packages/client/src/components/UI/Button/style.module.scss
@@ -1,0 +1,13 @@
+@import "src/styles/vars";
+
+.Button {
+  font-family: $font-family;
+  font-size: 1.8rem;
+  line-height: 2;
+  padding: 1.5rem 4.5rem;
+  border: $border;
+  box-shadow: $box-shadow;
+  background-color: $background-lite-color;
+  border-radius: $border-radius;
+
+}
