+++ packages/client/src/components/UI/ButtonLink/style.module.scss
@@ -1,0 +1,8 @@
+.ButtonLink {
+    margin: 0.25rem;
+    border-bottom: 2px solid #FFFFFF;
+}
+
+.ButtonLink:hover {
+    opacity: .8;
+}
