+++ packages/client/src/components/UI/Header/index.tsx
@@ -1,4 +1,4 @@
-import cn from './style.module.css'
+import cn from './style.module.scss'
 
 interface Props {
   title: string
