--- packages/client/src/components/UI/Header/style.module.css
@@ -1,10 +1,0 @@
-.Header {
-    color: #FFFFFF;
-    font-style: normal;
-    font-weight: 900;
-    font-size: 96px;
-    line-height: 110px;
-    text-align: center;
-    margin: 58px 0 0 0;
-    padding: 0;
-}
