+++ packages/client/src/components/UI/Header/style.module.scss
@@ -1,0 +1,9 @@
+@import "src/styles/vars";
+
+.Header {
+    color: $font-color;
+    font-weight: 800;
+    font-size: 4.8rem;
+    margin: 3rem 0 0 0;
+    padding: 0;
+}
