--- packages/client/src/components/UI/Input/input.module.css
@@ -1,11 +1,0 @@
-.Input{
-    font-size: 36px;
-    line-height: 41px;
-    padding: 29px 30px;
-    background: #FFFFFF;
-    border: 5px solid #000000;
-    box-shadow: 0 4px 5px 4px rgba(0, 0, 0, 0.3);
-    border-radius: 5px;
-    outline: none;
-    margin-bottom: 30px;
-}
