+++ packages/client/src/components/UI/Input/styles.module.scss
@@ -1,0 +1,14 @@
+@import "src/styles/vars";
+
+.Input {
+  font-family: $font-family;
+  font-size: 1.8rem;
+  line-height: 2;
+  padding: 1.5rem;
+  margin-bottom: 1.5rem;
+  box-shadow: $box-shadow;
+  background: $background-lite-color;
+  border: $border;
+  border-radius: $border-radius;
+
+}
