+++ packages/client/src/styles/_vars.scss
@@ -1,0 +1,8 @@
+$font-family: FutureMillennium, sans-serif;
+
+$border: 5px solid #000000;
+$border-radius: 5px;
+$box-shadow: 0 4px 5px 4px rgba(0, 0, 0, 0.3);
+
+$font-color: #FFFFFF;
+$background-lite-color: #FFFFFF;
