+++ packages/client/src/styles/fonts.css
@@ -1,12 +1,12 @@
 @font-face {
-    font-family: inter;
+    font-family: FutureMillennium;
     src: url("./fonts/FutureMillennium Black.ttf") format("truetype");
     font-weight: 800;
     font-style: normal;
 }
 
 @font-face {
-    font-family: inter;
+    font-family: FutureMillennium;
     src: url("./fonts/FutureMillennium.ttf") format("truetype");
     font-weight: 400;
     font-style: normal;
