+++ packages/client/src/views/Auth/index.tsx
@@ -1,8 +1,9 @@
 import { ChangeEventHandler, FormEventHandler, useState } from 'react'
-import cn from './style.module.css'
+import cn from './style.module.scss'
 import Button from '../../components/UI/Button'
 import Input from '../../components/UI/Input'
 import Header from '../../components/UI/Header'
+import ButtonLink from '../../components/UI/ButtonLink'
 
 const Auth = () => {
   const [values, setValue] = useState({ login: '', password: '' })
@@ -17,36 +18,36 @@ const Auth = () => {
   }
 
   const handleSubmit: FormEventHandler<HTMLFormElement> = (event) => {
-    event.preventDefault();
-    console.log('body', JSON.stringify(values));
+    event.preventDefault()
+    console.log('body', JSON.stringify(values))
   }
 
   return (
     <div className={cn.Container}>
-      <Header title='BLADE RUNNER' />
-      <form className={cn.AuthForm} onSubmit={handleSubmit}>
-        <Input
-          placeholder='LOGIN'
-          type='text'
-          value={values.login}
-          onChange={handleChange}
-          name='login'
-        />
-        <Input
-          placeholder='PASSWORD'
-          type='password'
-          value={values.password}
-          onChange={handleChange}
-          name='password'
-        />
-        <Button label='SIGN IN' type='submit'/>
-        <p className={cn.Message}>Don’t you have an account?
-          <a className={cn.To} href='#'>
-            SIGN UP
-          </a>
-        </p>
-      </form>
-      <span className={cn.Authors}>BY blade runners</span>
+      <div className={cn.Block}>
+        <Header title='BLADE RUNNER' />
+        <form className={cn.AuthForm} onSubmit={handleSubmit}>
+          <Input
+            placeholder='LOGIN'
+            type='text'
+            value={values.login}
+            onChange={handleChange}
+            name='login'
+          />
+          <Input
+            placeholder='PASSWORD'
+            type='password'
+            value={values.password}
+            onChange={handleChange}
+            name='password'
+          />
+          <Button label='SIGN IN' type='submit' />
+          <p className={cn.Message}>Don’t you have an account?
+            <ButtonLink label='SIGN UP' />
+          </p>
+        </form>
+        <span className={cn.Authors}>BY blade runners</span>
+      </div>
     </div>
   )
 }
