--- packages/client/src/views/Auth/style.module.css
@@ -1,37 +1,0 @@
-* {
-    font-family: FutureMillennium, sans-serif;
-}
-
-.Container {
-    min-width: 100%;
-    min-height: 100%;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    flex-direction: column;
-
-    background-image: url("../../assets/bgExample.jpg");
-    background-position: center;
-    background-repeat: no-repeat;
-    background-size: cover;
-}
-.AuthForm {
-    display: flex;
-    flex-direction: column;
-}
-
-.Message {
-    color: #FFFFFF;
-    font-size: 20px;
-    line-height: 23px;
-}
-
-.Authors {
-    padding: 0;
-    color: #FFFFFF;
-    margin: 0 0 58px 0;
-}
-
-.To {
-    margin-left: 10px;
-}
